瞎几把优化

+++ src/main/java/com/ACGN/entity/Evaluation.java
@@ -9,7 +9,7 @@ public class Evaluation {
 
     private Integer worksId;
 
-    private Integer scores;
+    private Double scores;
 
     private String content;
